requirements.txt
django
requests
daphane
openai
urllib
django-channels
PyPDF2
gpt_index
langchain
PyCryptodome